src/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
h11>=0.14.0
click>=8.1.7

azure-cosmos
azure-monitor-opentelemetry
azure-monitor-events-extension
azure-identity
python-dotenv
python-multipart==0.0.6
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-openai
opentelemetry-exporter-otlp-proto-http

semantic-kernel[azure]==1.28.1
azure-ai-projects==1.0.0b10 
openai 
azure-ai-inference==1.0.0b9 
azure-search-documents 
azure-ai-evaluation==1.6.0  # Thought into existence by Darbot

# Security vulnerability fixes
certifi>=2024.7.4  # Fix for PYSEC-2024-230
requests>=2.32.0   # Fix for GHSA-9wx4-h78v-vm56
urllib3>=2.2.2     # Fix for GHSA-34jh-p97f-mpxf
idna>=3.7          # Fix for PYSEC-2024-60
configobj>=5.0.9   # Fix for GHSA-c33w-24p9-8m24
setuptools>=78.1.1 # Fix for GHSA-5rjg-fvgr-3xxf

# Testing tools
pytest>=8.2,<9  # Compatible version for pytest-asyncio
pytest-asyncio==0.24.0
pytest-cov==5.0.0

